requirements.txt
mcp
fastmcp
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-hyperbrowser
requests
python-dotenv
